+++ test/async/service/supervised.rb
@@ -66,15 +66,17 @@ describe Async::Service::Supervisor::Supervised do
 			include Async::Service::Supervisor::Supervised
 		end
 		
-		state = {
-			name: "simple-service",
+		additional_state = {
 			endpoint: {name: "http", addresses: [{address: "127.0.0.1", port: 9292}]},
 		}
-		worker = environment.evaluator.supervisor_worker(state: state)
+		worker = environment.evaluator.supervisor_worker(state: additional_state)
 		worker_task = worker.run
 		
 		event = registration_monitor.pop
-		expect(event.supervisor_controller.state).to be == state
+		expect(event.supervisor_controller.state).to be == {
+			name: "simple-service",
+			endpoint: additional_state[:endpoint],
+		}
 	ensure
 		worker_task&.stop
 	end